Break down StoreTests into parts.
